requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.17.0
numpy<2.0.0
pandas
plotly
Pillow